# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DogDex API - Pokemon GO for dogs
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get current user's stored profile
      description: Returns the authenticated user's profile from the database
      responses:
        "200":
          description: User profile (null if not yet synced)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMyAccountData
      tags: [users]
      summary: Delete current user's Doggo Dex data
      description: >
        Permanently deletes the authenticated user's profile, leaderboard
        record, generated badge image records, generated badge files, and
        Clerk account.
      responses:
        "200":
          description: User data deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Account data deletion failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/sync:
    post:
      operationId: syncUser
      tags: [users]
      summary: Sync user profile after registration
      description: Creates or updates the user's profile with country and username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncUserRequest"
      responses:
        "200":
          description: User synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Database error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/collect:
    post:
      operationId: recordCollection
      tags: [users]
      summary: Record a new dog collection
      description: Increments the user's collection count and XP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectRequest"
      responses:
        "200":
          description: Collection recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Database error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get ranked users
      description: Returns leaderboard rankings. Filter by country or get global rankings.
      parameters:
        - name: country
          in: query
          schema:
            type: string
          description: Filter to a specific country (e.g. "United States")
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Max results to return
      responses:
        "200":
          description: Leaderboard rankings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
        "500":
          description: Database error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    MyProfile:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
          nullable: true
        country:
          type: string
        countryFlag:
          type: string
        collectionCount:
          type: integer
        xp:
          type: integer
        avatarUrl:
          type: string
          nullable: true
      required:
        - username
        - country
        - countryFlag
        - collectionCount
        - xp
    SyncUserRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 50
        displayName:
          type: string
          maxLength: 100
        country:
          type: string
          maxLength: 100
        countryFlag:
          type: string
          maxLength: 10
        avatarUrl:
          type: string
          format: uri
      required:
        - username
        - country
    CollectRequest:
      type: object
      properties:
        xpDelta:
          type: integer
          default: 10
          minimum: 0
          maximum: 1000
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
          nullable: true
        country:
          type: string
        countryFlag:
          type: string
        collectionCount:
          type: integer
        xp:
          type: integer
        avatarUrl:
          type: string
          nullable: true
      required:
        - rank
        - clerkId
        - username
        - country
        - countryFlag
        - collectionCount
        - xp
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message